+++ deepspace/src/deepspace/TestP4.java
@@ -152,6 +152,104 @@ public class TestP4 {
         System.out.println("State of the city: " + test_city);
         System.out.println("");
     }
+ 
+    static void testPowerEfficientStation(){
+        System.out.println("Testing PowerEfficientSpaceStation");
+        System.out.println("================================================================================");
+
+        SuppliesPackage base = new SuppliesPackage(1.1111f, 2.222f, 3.333f);
+        SpaceStation test_station = new SpaceStation("sergio", base);
+        PowerEfficientSpaceStation test_power = new PowerEfficientSpaceStation(test_station);
+        System.out.print("The created power efficient station is: " + test_power);
+        System.out.println("");
+
+        System.out.println("The station fires: " + test_power.fire());
+        System.out.println("The station protects: " + test_power.protection());
+        System.out.println("State of the power efficient station: " + test_power);
+        System.out.println("");
+
+        // BUG -- Lanza warnings que no deberia
+        Loot test_loot = new Loot(1, 2, 3, 4, 5, true, true);
+        System.out.println("The station receives the loot " + test_loot);
+        Transformation test_transform = test_power.setLoot(test_loot);
+        System.out.println("Transformation returned: " + test_transform);
+        System.out.println("State of the power efficient station: " + test_power);
+        
+    }
+
+    static void testBetaPowerEfficientStation(){
+        System.out.println("Testing BetaPowerEfficientSpaceStation");
+        System.out.println("================================================================================");
+        
+        SuppliesPackage base = new SuppliesPackage(1.1111f, 2.222f, 3.333f);
+        SpaceStation test_station = new SpaceStation("sergio", base);
+        BetaPowerEfficientSpaceStation test_beta = new BetaPowerEfficientSpaceStation(test_station);
+        System.out.println("The created BetaPowerEfficientSpaceStation is " + test_beta);
+
+        System.out.println("Firing a few times");
+        for(int i = 0; i < 5; i++){
+            System.out.println("\nThe station fires: " + test_beta.fire());
+        }
+        System.out.println("");
+
+        System.out.println("The Ui version of the station is " + test_beta.getUIversion());
+        System.out.println("");
+    }
+
+    static void testLoot(){
+        System.out.println("Testing Loot");
+        System.out.println("================================================================================");
+
+        Loot test_loot = new Loot(1, 2, 3, 4, 5, false, false);
+        System.out.println("The created loot is: " + test_loot);
+        System.out.println("");
+
+        System.out.println("The UI version is: " + test_loot.getUIversion());
+        System.out.println("");
+
+    }
+
+    static void testGameUniverse(){
+        System.out.println("Testing GameUniverse");
+        System.out.println("================================================================================");
+    
+        GameUniverse test_universe = new GameUniverse();
+        System.out.println("The created universe is " + test_universe);
+        System.out.println("");
+
+        System.out.println("Initializing the game with {\"Sergio\", \"Miguel\"}");
+        ArrayList<String> names = new ArrayList<>(0);
+        names.add("Sergio");
+        names.add("Miguel");
+        test_universe.init(names);
+        System.out.println("State after initializing: " + test_universe);
+        System.out.println("");
+
+        System.out.println("Combat");
+        CombatResult test_result = test_universe.combat();
+        System.out.println("Result of the combat: " + test_result);
+        System.out.println("State of the GameUniverse: " + test_universe);
+        System.out.println("");
+
+        System.out.println("Next turn:");
+        test_universe.nextTurn();
+        System.out.println("State of the GameUniverse: " + test_universe);
+        System.out.println("");
+
+        System.out.println("Combat");
+        test_result = test_universe.combat();
+        System.out.println("Result of the combat: " + test_result);
+        System.out.println("State of the GameUniverse: " + test_universe);
+        System.out.println("");
+
+        System.out.println("Next turn:");
+        test_universe.nextTurn();
+        System.out.println("State of the GameUniverse: " + test_universe);
+        System.out.println("");
+
+        System.out.println("Have a winner: " + test_universe.haveAWinner());
+
+    }
 
     public static void main(String[] args){
         //testNumericDamage();
@@ -159,12 +257,34 @@ public class TestP4 {
         System.out.println("");
         System.out.println("");
 
-        testSpecificDamage();
+        //testSpecificDamage();
         System.out.println("");
         System.out.println("");
         System.out.println("");
 
+        // Bug -- No se añade el loot de forma correcta
         //testSpaceCity();
+        System.out.println("");
+        System.out.println("");
+        System.out.println("");
+
+        // BUG -- El setLoot lanza warnings que no deberia
+        //testPowerEfficientStation();
+        System.out.println("");
+        System.out.println("");
+        System.out.println("");
+        
+        //testBetaPowerEfficientStation();
+        System.out.println("");
+        System.out.println("");
+        System.out.println("");
+
+        //testLoot();
+        System.out.println("");
+        System.out.println("");
+        System.out.println("");
+
+        testGameUniverse();
         System.out.println("");
         System.out.println("");
         System.out.println("");